refactoring -2

+++ src/video_converter.py
@@ -2,17 +2,20 @@ import os
 import shlex
 import subprocess
 import time
+from datetime import timedelta
 
-from src import logger
+from src import logger, write_to_csv_file
 
-# path = '/media/funker/3/FOTO/2017/2017-Samsung A5 and S7 Denys/2017-04/'
-path = '/Users/dstoianov/Documents/convert-video/'
+# path = '/Users/dstoianov/Documents/convert-video/'
+path = '/media/funker/3/FOTO/2014/'
 
-csv_file_name = 'movies_all_mp4_2017.csv'
+csv_file_name = f"{path.replace('/', '_').replace(' ', '_')}.csv"
 
 files = []
 files_ext = {}
-skipped_file_extensions = ['nfo', 'ini', 'jpg', 'txt', 'db', 'png', 'jpeg', 'sh']
+skipped_file_extensions = ['nfo', 'ini', 'jpg', 'nef', 'txt', 'db', 'png', 'jpeg', 'sh', 'gif', 'pdf', 'ppt', 'mp3',
+                           'xls',
+                           'mht', 'htm', 'zip', ]
 
 """
 http://coderunner.io/shrink-videos-with-ffmpeg-and-preserve-metadata/
@@ -58,6 +61,8 @@ def get_media_properties(filename):
 
 
 def read_files():
+    logger.info("=== " * 20)
+    logger.info("Collecting files..")
     local_files = []
     # r=root, d=directories, f = files
     for r, d, ff in os.walk(path):
@@ -72,29 +77,31 @@ def read_files():
             local_files.append({'name': file, 'size': size_mb, 'path': full_path})
 
     logger.info("Total collected '%s' files " % str(len(local_files)))
-    logger.info("=== " * 20)
     return local_files
 
 
 def convert_videos(files):
+    logger.info("=== " * 20)
     logger.info("Going to convert files..")
     # 18 more quality, 23 - default, 28 less quality
     crf = 23
-    prefix = '-crf-'
-    for filename in files:
+    prefix = '-ffmpeg-crf-'
+    g_start = time.time()
+    for i, filename in enumerate(files, start=1):
         if prefix in filename['name']:
-            continue
+            continue  # skipp already decoded files by prefix
         input_file = filename['path']
         chunks = input_file.split(".")
         output_file = f"{chunks[0]}{prefix}{crf}.mp4"
-        command = f"ffmpeg -i {input_file} -copy_unknown -map_metadata 0 -map 0 -codec copy \
+        command = f"ffmpeg -i \"{input_file}\" -copy_unknown -map_metadata 0 -map 0 -codec copy \
             -codec:v libx264 -pix_fmt yuv420p -crf {crf} \
             -codec:a aac -vbr 5 \
-            -preset medium {output_file}"
+            -preset medium \"{output_file}\""
         #   presets: slow, medium, fast
         # print(command)
 
-        logger.info("Converting file '%s - %sMb'..", filename['name'], filename['size'])
+        logger.info("%s) Converting file '%s - %sMB - %s sec'..", str(i), filename['name'], filename['size'],
+                    filename['duration'])
         start = time.time()
         if not os.path.isfile(output_file):
             res = subprocess.run(shlex.split(command), stdout=subprocess.PIPE, stderr=subprocess.STDOUT,
@@ -103,14 +110,16 @@ def convert_videos(files):
                 logger.error(res.stdout)
                 raise Exception('Error. Check parameters for decoding.')
 
-        end = time.time()
         size_mb = round(os.path.getsize(output_file) / 1024 / 1024, 2)
-        logger.info("\tElapsed time '%.4s' sec", end - start)
-        logger.info("\tOld size '%sMb' new size '%sMb', ratio '%.4s'", filename['size'], size_mb,
+        logger.info("\tElapsed time '%s'", time.strftime("%H:%M:%S", time.gmtime(time.time() - start)))
+        logger.info("\tOld size '%sMb' --> new size '%sMB', ratio '%.4s'", filename['size'], size_mb,
                     filename['size'] / size_mb)
+
+    logger.info("Total elapsed time for converting '%s'", str(timedelta(seconds=(time.time() - g_start))))
 
 
 def collect_statistics():
+    logger.info("=== " * 20)
     logger.info("Collect statistics..")
     file_types = []
     for file in files:
@@ -127,55 +136,61 @@ def collect_statistics():
     names = list(a['name'] for a in files)
     duplicates = list_duplicates(names)
     logger.info("Found '%s' duplicates %s", len(duplicates), duplicates)
-    logger.info("=== " * 20)
 
 
 def delete_broken_files(delete=False):
+    logger.info("=== " * 20)
+    logger.info("Delete broken files..")
     for file in files:
         if file.get('duration') is None:
-            logger.info("delete file '%s', size '%.5s' Mb", file['path'], file['size'])
+            logger.info("delete file '%s', size '%.5s' MB", file['path'], file['size'])
             if delete:
                 os.remove(file['path'])
             else:
                 logger.warn("Skipp deleting...")
-    logger.info("=== " * 20)
 
 
 def read_files_metadata():
+    logger.info("=== " * 20)
     logger.info("Collect file metadata..")
-    for i, file in enumerate(files):
-        logger.info("Processing %s) '%s' file..", str(i), file['name'])
+    for i, file in enumerate(files, start=1):
+        logger.info(" %s) Processing '%s' file..", str(i), file['name'])
         properties = get_media_properties(file['path'])
         file.update(properties)
 
-    logger.info("=== " * 20)
+    logger.info("======= Total collected '%s' files =======", len(files))
 
 
 def list_duplicates(seq):
     return sorted(set([x for x in seq if seq.count(x) > 1]))
 
 
-def delete_files(files: list):
-    logger.info("Potential released size '%sMb'", sum([i['size'] for i in files]))
+def delete_files(files: list, delete=False):
+    logger.info("Potential released size '%.7sMB'", sum([i['size'] for i in files]))
 
     for file in files:
         logger.info(f"Removing file '{file['path']}'..")
-        os.remove(file['path'])
+        if delete:
+            os.remove(file['path'])
+            # send2trash(file['path'])
+
+
+def delete_decoded_files(delete):
+    all_files = read_files()
+    files_to_delete = list(filter(lambda d: '-crf-23.mp4' not in d['name'], all_files))
+    delete_files(files_to_delete, delete)
 
 
 if __name__ == "__main__":
-    # files = read_files()
-    # collect_statistics()
+    files = read_files()
+    collect_statistics()
 
-    # read_files_metadata()
-    # delete_broken_files(delete=True)
-    # fnames = ['name', 'size', 'duration', 'mime_type', 'width', 'height', 'creation_date', 'last_modification', 'path']
-    # write_to_csv_file(csv_file_name, files, fnames)
-    # convert_videos(files)
+    read_files_metadata()
+    delete_broken_files(delete=False)
+    fnames = ['name', 'size', 'duration', 'mime_type', 'width', 'height', 'creation_date', 'last_modification', 'path']
+    write_to_csv_file(csv_file_name, files, fnames)
+    convert_videos(files)
 
-    all_files = read_files()
-
-    files_to_delete = list(filter(lambda d: '-crf-23.mp4' in d['name'], all_files))
-    delete_files(files_to_delete)
+    # delete_decoded_files(delete=False)
 
     logger.info('Done')
